Expanded comments and added optional code
Added optional code to verify that the stored data scope is local, and to check the data type of the input value for learning purposes. I've placed comment tags to the optional code to disable it. Also I've clarified some of the code explanation comments and added some additional comments.